requirements.txt
albumentations>=1.4.2
av>=11.0.0
black>=22.3.0
certifi>=2023.11.17
cityscapesscripts>=2.2.2
clip>=0.2.0
cloudpickle>=2.2.0
colour-science
comet_ml>=3.39.1
coremltools>=7.1
dask>=2022.9.2
dill>=0.3.6
docutils>=0.20.1
dvclive>=3.45.0
einops>=0.7.0
english_words>=2.0.1
evaluate>=0.4.1
fairscale>=0.4.12
filelock>=3.8.0
Flask>=3.0.2
fvcore>=0.1.5.post20220512
GitPython>=3.1.42
h5py>=3.10.0
hydra-core>=1.3.2
imageio>=2.22.0
imagesize>=1.4.1
imutils>=0.5.4
iopath>=0.1.9
ipython>=8.12.3
joblib>=1.2.0
labelbox>=3.34.0
langchain>=0.1.12
matplotlib>=3.6.0
memory_profiler>=0.61.0
mmdet>=3.3.0
mss>=9.0.1
nltk>=3.8.1
nncf>=2.9.0
numpy>=1.24.3
omegaconf>=2.2.3
# onnx>=1.14.0
# onnxruntime>=1.15.0
# onnxsim>=0.4.36
openai>=0.27.5
opencv_contrib_python>=4.8.0.76
opencv_python>=4.8.0.76
openpyxl>=3.1.2
openvino>=2024.0.0
pafy>=0.5.5
pandas>=1.5.0
pandastable>=0.13.1
parse>=1.20.1
Pillow>=10.2.0
protobuf>=3.20.0#"protobuf>=3.21,<5.0.0"#protobuf>=3.20.0
psutil>=5.9.4
py_cpuinfo>=9.0.0
pyexiv2>=2.8.1
Pygments>=2.13.0
pytest>=8.1.1
PyYAML>=6.0.1
rapidfuzz>=2.13.7
rawpy>=0.18.1
ray>=2.9.3
Requests>=2.31.0
rich>=13.7.1
scikit_image==0.24.0
scikit_learn==1.5.2
scipy==1.14.1
seaborn>=0.13.2
sentence_transformers>=2.2.2
sentry_sdk>=1.9.10
setuptools>=58.1.0
Shapely>=1.8.5.post1
shortuuid>=1.0.13
Sphinx>=7.2.6
sqlparse>=0.4.4
statsmodels>=0.14.0
streamlit>=1.27.2
tabulate>=0.9.0
tensorflow>=2.16.1
termcolor
thinc>=8.2.3
thop>=0.1.1.post2209072238
timm>=0.6.11
tqdm>=4.65.0
transformers>=4.30.0.dev0
urllib3>=1.26.12
visdom>=0.2.4
vit_pytorch>=0.37.1
wandb>=0.13.4
x2paddle>=1.4.1
xmltodict>=0.13.0
yt_dlp>=2024.3.10
beautifulsoup4>=4.12.3
opencv-contrib-python>=4.9.0.80 # 4.8.0.76
selenium==4.21.0
undetected-chromedriver==3.5.5
aiohttp_retry
requests
labelbox
shapely
aiofiles
streamlit-drawable-canvas
paramiko>=3.5.0
cryptography>=43.0.1
ete3 
biopython
ffmpeg-python
ect
phate
scprep
dask[dataframe]
umap-learn
scikit-learn-extra


dask==2024.9.0
distributed==2024.9.0
dask-expr==1.1.14
fsspec[http]==2024.3.1
datasets==2.19.1
rapids-dask-dependency==24.10.0